Hooks cleaned
