+++ src/hooks/index.ts
@@ -1,0 +1,2 @@
+export { default as useTheme } from './useTheme/useTheme'
+export { default as useOutsideClick } from './useOutsideClick/useOutsideClick'
